+++ assets/src/components/activities/short_answer/ShortAnswerAuthoring.tsx
@@ -1,33 +1,33 @@
+import { AuthoringButtonConnected } from 'components/activities/common/authoring/AuthoringButton';
+import { InputTypeDropdown } from 'components/activities/common/authoring/InputTypeDropdown';
+import { ActivitySettings } from 'components/activities/common/authoring/settings/ActivitySettings';
+import { shuffleAnswerChoiceSetting } from 'components/activities/common/authoring/settings/activitySettingsActions';
+import { Hints } from 'components/activities/common/hints/authoring/HintsAuthoringConnected';
+import { ResponseActions } from 'components/activities/common/responses/responseActions';
+import { ResponseCard } from 'components/activities/common/responses/ResponseCard';
+import { SimpleFeedback } from 'components/activities/common/responses/SimpleFeedback';
+import { Stem } from 'components/activities/common/stem/authoring/StemAuthoringConnected';
+import { StemDelivery } from 'components/activities/common/stem/delivery/StemDelivery';
+import { DEFAULT_PART_ID } from 'components/activities/common/utils';
+import { InputEntry } from 'components/activities/short_answer/sections/InputEntry';
+import { getTargetedResponses, shortAnswerOptions } from 'components/activities/short_answer/utils';
+import { makeResponse, Manifest, Response } from 'components/activities/types';
+import { TabbedNavigation } from 'components/tabbed_navigation/Tabs';
+import { getCorrectResponse } from 'data/activities/model/responses';
+import { containsRule, eqRule } from 'data/activities/model/rules';
+import { defaultWriterContext } from 'data/content/writers/context';
 import React from 'react';
 import ReactDOM from 'react-dom';
+import { Provider } from 'react-redux';
+import { configureStore } from 'state/store';
 import {
   AuthoringElement,
   AuthoringElementProps,
   AuthoringElementProvider,
   useAuthoringElementContext,
 } from '../AuthoringElement';
-import { ShortAnswerModelSchema } from './schema';
-import * as ActivityTypes from '../types';
 import { ShortAnswerActions } from './actions';
-import { ModalDisplay } from 'components/modal/ModalDisplay';
-import { Provider } from 'react-redux';
-import { configureStore } from 'state/store';
-import { TabbedNavigation } from 'components/tabbed_navigation/Tabs';
-import { Hints } from 'components/activities/common/hints/authoring/HintsAuthoringConnected';
-import { StemDelivery } from 'components/activities/common/stem/delivery/StemDelivery';
-import { defaultWriterContext } from 'data/content/writers/context';
-import { parseInputFromRule } from 'components/activities/common/responses/authoring/rules';
-import { Stem } from 'components/activities/common/stem/authoring/StemAuthoringConnected';
-import { SimpleFeedback } from 'components/activities/common/responses/SimpleFeedback';
-import { getCorrectResponse } from 'components/activities/common/responses/authoring/responseUtils';
-import { ResponseActions } from 'components/activities/common/responses/responseActions';
-import { ActivitySettings } from 'components/activities/common/authoring/settings/ActivitySettings';
-import { shuffleAnswerChoiceSetting } from 'components/activities/common/authoring/settings/activitySettingsActions';
-import { InputTypeDropdown } from 'components/activities/short_answer/sections/InputTypeDropdown';
-import { AuthoringButtonConnected } from 'components/activities/common/authoring/AuthoringButton';
-import { InputEntry } from 'components/activities/short_answer/sections/InputEntry';
-import { getTargetedResponses } from 'components/activities/short_answer/utils';
-import { ResponseCard } from 'components/activities/common/responses/ResponseCard';
+import { ShortAnswerModelSchema } from './schema';
 
 const store = configureStore();
 
@@ -40,15 +40,11 @@ const ShortAnswer = () => {
           <div className="d-flex flex-column flex-md-row mb-2">
             <Stem />
             <InputTypeDropdown
+              options={shortAnswerOptions}
               editMode={editMode}
-              inputType={model.inputType}
+              selected={model.inputType}
               onChange={(inputType) =>
-                dispatch(
-                  ShortAnswerActions.setInputType(
-                    inputType,
-                    parseInputFromRule(getCorrectResponse(model).rule),
-                  ),
-                )
+                dispatch(ShortAnswerActions.setInputType(inputType, DEFAULT_PART_ID))
               }
             />
           </div>
@@ -57,20 +53,20 @@ const ShortAnswer = () => {
           <div className="d-flex flex-column mb-2">
             <StemDelivery stem={model.stem} context={defaultWriterContext()} />
             <InputEntry
-              key={getCorrectResponse(model).id}
+              key={getCorrectResponse(model, DEFAULT_PART_ID).id}
               inputType={model.inputType}
-              response={getCorrectResponse(model)}
+              response={getCorrectResponse(model, DEFAULT_PART_ID)}
               onEditResponseRule={(id, rule) => dispatch(ResponseActions.editRule(id, rule))}
             />
-            <SimpleFeedback />
-            {getTargetedResponses(model).map((response: ActivityTypes.Response) => (
+            <SimpleFeedback partId={DEFAULT_PART_ID} />
+            {getTargetedResponses(model, DEFAULT_PART_ID).map((response: Response) => (
               <ResponseCard
                 title="Targeted feedback"
                 response={response}
                 updateFeedback={(id, content) =>
                   dispatch(ResponseActions.editResponseFeedback(response.id, content))
                 }
-                onRemove={(id) => dispatch(ResponseActions.removeResponse(id))}
+                removeResponse={(id) => dispatch(ResponseActions.removeResponse(id))}
                 key={response.id}
               >
                 <InputEntry
@@ -83,7 +79,18 @@ const ShortAnswer = () => {
             ))}
             <AuthoringButtonConnected
               className="align-self-start btn btn-link"
-              action={() => dispatch(ShortAnswerActions.addResponse())}
+              action={() =>
+                dispatch(
+                  ResponseActions.addResponse(
+                    makeResponse(
+                      model.inputType === 'numeric' ? eqRule('1') : containsRule('another answer'),
+                      0,
+                      '',
+                    ),
+                    DEFAULT_PART_ID,
+                  ),
+                )
+              }
             >
               Add targeted feedback
             </AuthoringButtonConnected>
@@ -91,7 +98,7 @@ const ShortAnswer = () => {
         </TabbedNavigation.Tab>
 
         <TabbedNavigation.Tab label="Hints">
-          <Hints hintsPath="$.authoring.parts[0].hints" />
+          <Hints partId={DEFAULT_PART_ID} />
         </TabbedNavigation.Tab>
         <ActivitySettings settings={[shuffleAnswerChoiceSetting(model, dispatch)]} />
       </TabbedNavigation.Tabs>
@@ -112,5 +119,5 @@ export class ShortAnswerAuthoring extends AuthoringElement<ShortAnswerModelSchem
   }
 }
 // eslint-disable-next-line
-const manifest = require('./manifest.json') as ActivityTypes.Manifest;
+const manifest = require('./manifest.json') as Manifest;
 window.customElements.define(manifest.authoring.element, ShortAnswerAuthoring);